requirements.txt
blaseball-mike
Pillow
requests